Updating to sites to watch.
These changes to @rodericj code will abandon the by payload transfer in favor of a smaller amount of data transfer. I also added back complication creation on the watch, but the Network code needs to be rewritten to be more efficient on the watch so that the data can update efficiently.